+++ views/okta_webhook.py
@@ -4,6 +4,7 @@ import os
 import json
 import glob
 import time
+from collections import defaultdict
 from time import sleep
 import requests
 import re
@@ -12,24 +13,44 @@ from flask import (Flask, request, render_template,
                    session, redirect, url_for, escape,
                    send_from_directory, Blueprint, abort)
 
-new_okta = Blueprint('okta_new', __name__)
+from bin.view_modifiers import response
+
+blueprint = Blueprint('okta_webhook', __name__)
 
 server_json_file = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'data', 'server.json'))
 okta_json_file = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'data', 'okta_json.json'))
 okta_verification_file = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'bin', 'okta_verification.py'))
-jp_webhooks_file = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'data', 'webhooks.json'))
+webhooks_file = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'data', 'webhooks.json'))
 scripts_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'scripts'))
 
 
-@new_okta.route('/okta_new', methods=['GET', 'POST'])
+@blueprint.route('/webhooks/okta', methods=['GET', 'POST'])
+@response(template_file='webhooks/okta/home.html')
+def okta_webhook():
+    if not 'username' in session:
+        return redirect(url_for('logout'))
+    with open(webhooks_file, 'r') as fin:
+        webhooks_json = json.load(fin)
+    okta_webhooks_list = []
+    for each_webhook in webhooks_json:
+        data = defaultdict(lambda: "MISSING", each_webhook)
+        tag = data['tag']
+        if tag == "okta":
+            okta_webhooks_list.append(each_webhook)
+    print(okta_webhooks_list)
+
+    return {'key': "Have a good weekend my dude.", 'username': session.get('username'),
+            'okta_list': okta_webhooks_list}
+
+@blueprint.route('/webhooks/okta/new', methods=['GET', 'POST'])
 def okta_new():
     if 'username' in session:
         os.chmod(okta_verification_file, mode=0o0755)
         # okta_json = '/usr/local/jawa/okta_json.json'
 
-        if not os.path.isfile(jp_webhooks_file):
+        if not os.path.isfile(webhooks_file):
             data = []
-            with open(jp_webhooks_file, 'w') as outfile:
+            with open(webhooks_file, 'w') as outfile:
                 json.dump(data, outfile, indent=4)
 
         if request.method == 'POST':
@@ -43,8 +64,8 @@ def okta_new():
             with open(server_json_file) as json_file:
                 data = json.load(json_file)
                 server_address = data['jawa_address']
-            if not os.path.isdir('/usr/local/jawa/'):
-                os.mkdir('/usr/local/jawa/')
+            # if not os.path.isdir('/usr/local/jawa/'):
+            #     os.mkdir('/usr/local/jawa/')
 
             if not os.path.isdir(scripts_dir):
                 os.mkdir(scripts_dir)
@@ -65,8 +86,8 @@ def okta_new():
 
             old_script_file = os.path.join(scripts_dir, f.filename)
 
-            # hooks_file = '/etc/webhook.conf'
-            data = json.load(open(jp_webhooks_file))
+            # hooks_file = '/etc/webhooks.conf'
+            data = json.load(open(webhooks_file))
 
             new_id = okta_name
             script_file = os.path.join(scripts_dir, okta_name + "_" + f.filename)
@@ -74,7 +95,7 @@ def okta_new():
             os.rename(old_script_file, script_file)
             new_file = script_file
 
-            okta_info = json.load(open(jp_webhooks_file))
+            okta_info = json.load(open(webhooks_file))
 
             for i in data:
                 if str(i['name']) == okta_name:
@@ -134,7 +155,7 @@ def okta_new():
                               "webhook_password": "null",
                               "tag": "okta"})
 
-            with open(jp_webhooks_file, 'w') as outfile:
+            with open(webhooks_file, 'w') as outfile:
                 json.dump(okta_info, outfile, indent=4)
 
             # Verify/activate
@@ -152,7 +173,76 @@ def okta_new():
             return render_template('success.html', login="true")
 
         else:
-            return render_template('okta_new.html', setup="setup", username=str(escape(session['username'])))
+            return render_template('webhooks/okta/new.html', setup="setup",
+                                   username=str(escape(session['username'])))
 
     else:
         return render_template('home.html', login="false")
+
+
+@blueprint.route('/okta_delete', methods=['GET', 'POST'])
+def okta_delete():
+    exists = os.path.isfile(server_json_file)
+    if exists == False:
+        return render_template('setup.html',
+                               setup="setup",
+                               username=str(escape(session['username'])))
+    if 'username' not in session:
+        return redirect(url_for('logout'))
+
+    with open(webhooks_json, 'r+') as fin:
+        content = fin.read()
+    webhook_data = json.loads(content)
+    i = 0
+    names = []
+    for item in webhook_data:
+        data = defaultdict(lambda: "MISSING", item)
+        tag = item['tag']
+        if tag == "okta":
+            names.append(str(webhook_data[i]['name']))
+        i += 1
+
+    content = names
+
+    if request.method == 'POST':
+        if request.form.get('webhookname') != '':
+
+            webhookname = request.form.get('webhookname')
+            ts = time.time()
+
+            data = json.load(open(webhooks_json))
+
+            for d in data:
+                if d['name'] == webhookname:
+                    scriptPath = (d['script'])
+                    newScriptPath = scriptPath + '.old'
+                    if os.path.exists(scriptPath):
+                        os.rename(scriptPath, newScriptPath)
+
+            data[:] = [d for d in data if d.get('id') != webhookname]
+
+            with open(webhooks_json, 'w') as outfile:
+                json.dump(data, outfile)
+
+            data = json.load(open(webhooks_json))
+
+            for d in data:
+                if d['name'] == webhookname:
+                    response = requests.post(
+                        d['okta_url'] + '/api/v1/eventHooks/{}/lifecycle/deactivate'.format(d['okta_id']),
+                        headers={"Authorization": "SSWS {}".format(d['okta_token'])})
+
+                    response = requests.delete(d['okta_url'] + '/api/v1/eventHooks/{}'.format(d['okta_id']),
+                                               headers={"Authorization": "SSWS {}".format(d['okta_token'])})
+
+            data[:] = [d for d in data if d.get('name') != webhookname]
+
+            with open(webhooks_json, 'w') as outfile:
+                json.dump(data, outfile)
+
+        return redirect(url_for('success'))
+
+    else:
+        return render_template('okta_delete.html',
+                               text=content, delete="delete",
+                               username=str(escape(session['username'])))
